Refactor ShopifyBulkOperations to use a dedicated card component.
This PR simplifies `src/views/ShopifyBulkOperations.vue` by extracting its repeated operation-card presentation into a typed, focused component (`ShopifyBulkOperationCard.vue`). It computes stable display values once per operation instead of doing it during the render loop. Tests are added to cover rendering and event emission for the new component.

* Extracted markup to `src/components/ShopifyBulkOperationCard.vue`.
* Removed repetitive render functions in favor of a pure display-model helper (a single `display` computed property mapping the state).
* Preserved rendered structure, styling, ordering, and behavior precisely.
* Emitting `view-system-message` and `copy-id` back to the parent.
* Added `ShopifyBulkOperationCard.spec.ts` for unit test coverage.
* Introduced `src/types/ShopifyBulkOperation.ts` containing strict `ShopifyBulkOperation` and `HotwaxMessage` types to enforce tight contracts without polluting the store.
* Fixed test suite to use fully typed fixtures instead of bypassing TypeScript rules.
* Removed unused imports, icons, and duplicate CSS styles.
